Moved Extras.kt to core module
