+++ app/src/main/java/com/yashovardhan99/healersdiary/onboarding/OnboardingRepository.kt
@@ -1,7 +1,7 @@
 package com.yashovardhan99.healersdiary.onboarding
 
 import com.yashovardhan99.healersdiary.database.HealersDao
-import com.yashovardhan99.healersdiary.utils.DangerousDatabase
+import com.yashovardhan99.core.DangerousDatabase
 import javax.inject.Inject
 import javax.inject.Singleton
