+++ app/src/main/java/com/yashovardhan99/healersdiary/onboarding/OnboardingViewModel.kt
@@ -11,7 +11,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.yashovardhan99.healersdiary.AppDataStore
 import com.yashovardhan99.core.analytics.AnalyticsEvent
-import com.yashovardhan99.healersdiary.utils.DangerousDatabase
+import com.yashovardhan99.core.DangerousDatabase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.collect
