+++ app/src/main/java/com/yashovardhan99/healersdiary/patients/PatientDetailFragment.kt
@@ -16,6 +16,8 @@ import androidx.recyclerview.widget.ConcatAdapter
 import androidx.recyclerview.widget.GridLayoutManager
 import com.google.android.material.transition.MaterialContainerTransform
 import com.yashovardhan99.core.analytics.AnalyticsEvent
+import com.yashovardhan99.core.getColorFromAttr
+import com.yashovardhan99.core.transitionDurationLarge
 import com.yashovardhan99.healersdiary.R
 import com.yashovardhan99.healersdiary.dashboard.DashboardViewModel
 import com.yashovardhan99.healersdiary.database.ActivityType
